% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_tss_calculation.R
\name{construct_introns}
\alias{construct_introns}
\title{Construct introns from GRanges object}
\usage{
construct_introns(x, update = TRUE)
}
\arguments{
\item{x}{An object of class \code{GRanges}. It has to have \code{exon}
feature present.}

\item{update}{If \code{TRUE} (default), \code{x} is updated by reference
and returned invisibily, else just the \code{intron} coordinates are
returned.}
}
\value{
An object of class \code{GRanges} with updated \code{intron}
coordinates or just the intron coordinates depending on \code{update}.
They all inherit from \code{GRanges}.
}
\description{
This function generates intronic coordinates by extracting
all the \code{exons} from a \code{GRanges} object.
}
\examples{
gff.file = system.file('extdata/chr21.refseq.hg19.gtf', package='genomation')
gff = gffToGRanges(gff.file)
construct_introns(gff)
}
