% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_tss_calculation.R
\name{mixed_circ_tss_calculation}
\alias{mixed_circ_tss_calculation}
\title{Function for reading exon intron and promoter structure from a given bed or GFF file}
\usage{
readTranscriptFeatures(location,remove.unusual=TRUE,
                              up.flank=1000,down.flank=1000,unique.prom=TRUE)
}
\arguments{
\item{location}{location of the bed file with 12 or more columns.
The file can end in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip}
and/or start with \code{http://} or \code{ftp://}. If the file is not compressed
it can also start with \code{https://} or \code{ftps://}. or location of GFF file}

\item{gff}{If TRUE, location is assumed to be GFF file}

\item{id}{This is supposed to be a column of GFF file showing the symbol or id of each genes}

\item{remove.unusual}{remove the chromomesomes with unsual names, mainly random chromsomes etc (only in bed file)}

\item{up.flank}{up-stream from TSS to detect promoter boundaries}

\item{down.flank}{down-stream from TSS to detect promoter boundaries}

\item{unique.prom}{get only the unique promoters, promoter boundaries will not have
a gene name if you set this option to be TRUE}
}
\value{
a \code{\link{GRangesList}} containing locations of exon/intron/promoter/TSS
}
\description{
Function for reading exon intron and promoter structure from a given bed or GFF file
}
\details{
This function is an extension of readTranscriptFeatures function from genomation package that now supports GFF file.
The idea behind GFF features is that they are calculated on gene levels in contrast to transcript levels. Therefore the parameter id must be supplied
together with GFF file so that the function knows what genes to use. For each unique genes, the function collapse the coordinate to contain all the genomics features.
These new coordinates are used to calculate TSS and promoter regions.
}
\note{
one bed track per file is only accepted, the bed files with multiple tracks will cause en error.
}
\examples{
my.bed12.file = system.file("extdata/chr21.refseq.hg19.bed", package = "genomation")
my.bed12.file
feats = mixed_circ_tss_calculation(my.bed12.file)
names(feats)
sapply(feats, head)

gff.file = system.file('extdata/chr21.refseq.hg19.gtf', package='genomation')
feats = mixed_circ_tss_calculation(gff.file,gff=TRUE)
names(feats)

}
