% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_adjust.R
\name{mixedcirc_adjust}
\alias{mixedcirc_adjust}
\title{Perform p-value adjustment on mixedcirc_fit_list}
\usage{
mixedcirc_adjust(
  input,
  method = "BH",
  pattern = "p_value",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{input}{An object of mixedcirc_fit_list. It has to contain more than one variable}

\item{method}{A vector of character strings containing the names of the multiple testing procedures for which adjusted p-values are to be computed. This vector should include any of the following: "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD", "BH", "BY", "ABH", "TSBH".}

\item{pattern}{Columns of results with this pattern will be selected for doing adjustment. Default: p_value}

\item{verbose}{Show information about different stages of the processes. Default FALSE}

\item{...}{additional arguments passed to mt.rawp2adjp from multtest package}
}
\value{
A class of mixedcirc_fit_list in which the adjusted p-values have been added in the results
}
\description{
This functions performs p-value adjustment on mixedcirc_fit_list
}
\details{
The p-values from the matching columns will be adjusted one column at the time using the selected method.
}
\examples{
data("circa_data")

results<-mixedcirc_detect(data_input = circa_data$data_matrix,
time = circa_data$time,group = circa_data$group,id = circa_data$id,period = 24,verbose = TRUE)
results<-mixedcirc_adjust(results,method="Bonferroni")

}
