% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_detectpeaks.R
\name{mixedcirc_detectpeaks}
\alias{mixedcirc_detectpeaks}
\title{Finds peak of a rhythm}
\usage{
mixedcirc_detectpeaks(
  phi,
  period = 24,
  min_time = 0,
  correct_period = T,
  correct_min = F
)
}
\arguments{
\item{phi}{phase of a peak}

\item{period}{Period of circadian rhythm. Default: 24}

\item{min_time}{Minimum time considered}

\item{correct_period}{If TRUE, period will be forced to be between the min and max time}

\item{correct_min}{minimum time will be adjusted}
}
\value{
the peak time
}
\description{
This functions finds peak of a rhythm
}
\details{

}
\examples{
data("circa_data")

results <- mixedcirc_detect(data_input = circa_data$data_matrix,
time = circa_data$time,
group = circa_data$group,
id = circa_data$id,
period = 24,
verbose = TRUE)
mixedcirc_detectpeaks(results[1][,5],min_time=0)

}
