% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mixedcirc_fit-class}
\alias{mixedcirc_fit-class}
\alias{mixedcirc_fit}
\title{Represents the result of model fitting on a single variable.}
\description{
Represents the result of model fitting on a single variable.
}
\section{Slots}{

\describe{
\item{\code{results}}{A one-row data frame containing statistics from the fitted model.}

\item{\code{fit}}{The actual fitted model object.}

\item{\code{exp_design}}{The experimental design data frame created by mixedcirc_detect.}

\item{\code{type}}{The type of analysis performed, for example "RRBS".}

\item{\code{call}}{The function call used to create the object.}

\item{\code{params}}{A list of parameters used by other functions.}
}}

