% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mixedcirc_fit_list-class}
\alias{mixedcirc_fit_list-class}
\alias{mixedcirc_fit_list}
\title{Represents the result of model fitting on several variables.}
\description{
Represents the result of model fitting on several variables.
}
\section{Slots}{

\describe{
\item{\code{results}}{A list of mixedcirc_fit}
}}

