% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mixedcirc_integration-class}
\alias{mixedcirc_integration-class}
\alias{mixedcirc_integration}
\title{Represents the result of integration on multiple data sets}
\description{
Represents the result of integration on multiple data sets
}
\section{Slots}{

\describe{
\item{\code{partial}}{A list of partial scores for individual data sets}

\item{\code{average}}{A matrix of average scores}

\item{\code{loadings}}{A List of loadings for individual data sets}

\item{\code{model}}{The original model fitted}
}}

