% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.mixedcirc_fit_list}
\alias{print.mixedcirc_fit_list}
\title{Prints circadian rhythm}
\usage{
\method{print}{mixedcirc_fit_list}(x, ...)
}
\description{
A simple function for printing the mixedcirc_fit object
}
