% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,mixedcirc_fit-method}
\alias{show,mixedcirc_fit-method}
\title{Prints circadian rhythm}
\usage{
\S4method{show}{mixedcirc_fit}(object)
}
\description{
A simple function for printing the mixedcirc_fit object
}
