% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_rrbs_voom.R
\name{variance_fit}
\alias{variance_fit}
\title{This function is used to calculate sigma of fit.}
\usage{
variance_fit(
  exprObj,
  formula,
  data,
  REML = FALSE,
  useWeights = TRUE,
  weightsMatrix = NULL,
  showWarnings = TRUE,
  fxn = identity,
  colinearityCutoff = 0.999,
  control = lme4::lmerControl(calc.derivs = FALSE, check.rankX = "stop.deficient"),
  quiet = quiet,
  BPPARAM = bpparam(),
  ignore_na = FALSE,
  ...
)
}
\description{
This function is used to calculate sigma of fit.
}
