% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_detect.R
\name{mixedcirc_detect}
\alias{mixedcirc_detect}
\title{Perform differential circadian rhythm analysis using mixed models}
\usage{
mixedcirc_detect(
  data_input = NULL,
  time = NULL,
  group = NULL,
  id = NULL,
  period = 24,
  lm_method = c("lm", "lme")[2],
  f_test = c("multcomp_f", "multcomp_chi", "Satterthwaite", "Kenward-Roger")[3],
  abs_phase = TRUE,
  obs_weights = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data_input}{A numerical matrix or data.frame (N*P) or DGEList where in the rows are samples (N) and the columns are variables (P). If DGEList is provided, regression weights will be estimated!}

\item{time}{A vector of length N, showing circadian time of each sample}

\item{group}{A character vector of length N. If performing differential circadian rhythm analysis, group is a factor, showing grouping of the samples. Analysis of two groups is supported at this stage! See details!}

\item{id}{A vector of length N showing identity of each *unique* sample. See details}

\item{period}{Period of circadian rhythm. Default: 24}

\item{lm_method}{The regression method to use. At this stage, `lm` and `lme` are supported! If lm is selected, normal regression will be performed. Default: "lme"}

\item{f_test}{Type of f-test for calculating p-value of the rhythm. Possible values are "multcomp_f","multcomp_chi","Satterthwaite", "Kenward-Roger". Default: Satterthwaite}

\item{abs_phase}{Whether to return absolute phase or not. Default: TRUE}

\item{obs_weights}{Regression weights. Default: NULL. See details}

\item{verbose}{Show information about different stages of the processes. Default FALSE}

\item{...}{additionl arguments to the regression function}
}
\value{
A class of mixedcirc_fit_list.
}
\description{
This functions performs differential circadian rhythm analysis using mixed models.
}
\details{
For each variable we use the following mode:
In this part we do rhythmicity analysis on individual variables using the following model:
measure ~0 + group + group:in + group:out
Where `in` is defined as cos(2 \* pi \* time / period) and `out` as sin(2 \* pi \* exp_design$time / period).
The global inference is tested on group:in == 0 or group:out == 0
The between group difference is tested as differences between in and out across different groups. No p-value adjustment is performed!

`obs_weights` is a matrix of size N*P where each colum shows the weights for all the observations for that particular variable.
}
\examples{
data("circa_data")

results<-mixedcirc_detect(data_input = circa_data$data_matrix,
time = circa_data$time,group = circa_data$group,id = circa_data$id,period = 24,verbose = TRUE)

}
