% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_detect.R
\name{mixedcirc_detect}
\alias{mixedcirc_detect}
\title{Perform differential circadian rhythm analysis using mixed models}
\usage{
mixedcirc_detect(
  data_input = NULL,
  time = NULL,
  group = NULL,
  id = NULL,
  period = 24,
  lm_method = c("lm", "lme", "nlme")[2],
  f_test = c("multcomp_f", "multcomp_chi", "Satterthwaite", "Kenward-Roger")[3],
  abs_phase = TRUE,
  obs_weights = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data_input}{A numerical matrix or data.frame (N*P) where in the rows are samples (N) and the columns are variables (P)}

\item{time}{A vector of length N, showing circadian time of each sample}

\item{group}{A character vector of length N. If performing differential circadian rhythm analysis, group is a factor, showing grouping of the samples. Analysis of two groups is supported at this stage! See details!}

\item{id}{A vector of length N showing identity of each *unique* sample. See details}

\item{period}{Period of circadian rhythm. Default: 24}

\item{lm_method}{The regression method to use. At this stage, `lm`,`lme`,`nlme` are supported! If lm is selected, normal regression will be performed. Default: "lme"}

\item{f_test}{Type of f-test for calculating p-value of the rhythm. Possible values are "multcomp_f","multcomp_chi","Satterthwaite", "Kenward-Roger". Default: Satterthwaite}

\item{abs_phase}{Whether to return absolute phase or not. Default: TRUE}

\item{obs_weights}{Regression weights. Default: NULL. See details}

\item{verbose}{Show information about different stages of the processes. Default FALSE}

\item{...}{additionl arguments to the regression function}
}
\value{
A nested list. Each element of the first list contains a data frame (result) and a fit class. The data frame includes the estimated mesor, amplitude and phase for each of the groups.
The data frame also includes the p-values for each estimates as well as differences between the groups.
The fit object is the actual model fitted.
}
\description{
This functions performs differential circadian rhythm analysis using mixed models.
}
\details{

}
\examples{
data("circa_data")

results<-mixedcirc_detect(data_input = circa_data$data_matrix,
time = circa_data$time,group = circa_data$group,id = circa_data$id,period = 24,verbose = TRUE)

}
