% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mixedcirc_fit-class}
\alias{mixedcirc_fit-class}
\alias{mixedcirc_fit}
\title{Represents the result of model fitting on a single variable.}
\description{
Represents the result of model fitting on a single variable.
}
\section{Slots}{

\describe{
\item{\code{results}}{A one row data frame that contains statistics of fitted model}

\item{\code{fit}}{is the actual fitted model}

\item{\code{exp_design}}{is the experimental design file created by mixedcirc_detect}

\item{\code{type}}{represents the type of analysis done. E.g. RRBS etc}
}}

