% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_fit_plot.R
\name{mixedcirc_fit_plot}
\alias{mixedcirc_fit_plot}
\title{Plots circadian rhythm}
\usage{
mixedcirc_fit_plot(
  x,
  period = 24,
  min_time = NULL,
  max_time = NULL,
  plot_title = "variable",
  plot_points = TRUE,
  plot_smooth = FALSE,
  plot_trend = TRUE,
  xlab = "time",
  ylab = "Expression"
)
}
\arguments{
\item{x}{A class of mixedcirc_fit}

\item{period}{The rhythm period. Default: 24}

\item{min_time}{Minimum time span to do the prediction. If NULL, it will be taken from the fit. Default: NULL}

\item{max_time}{Maximum time span to do the prediction. If NULL, it will be taken from the fit. Default: NULL}

\item{plot_title}{Title of the plot.}

\item{plot_points}{whether to plot the individual points or not. Default TRUE}

\item{plot_smooth}{Whether to plot smoothing line or not. Default: FALSE}

\item{plot_trend}{Whether to plot trend lines or not. Default: TRUE}

\item{xlab}{Label on the x-axis}

\item{ylab}{Label on the y-axis}

\item{y}{Not used}
}
\value{
A ggplot object.
}
\description{
This functions plots circadian rhythm fitted using mixedcirc_detect function.
}
\examples{
data("circa_data")

results<-mixedcirc_detect(data_input = circa_data$data_matrix,
time = circa_data$time,group = circa_data$group,id = circa_data$id,period = 24,verbose = TRUE)
plot(results[[1]])
}
