% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_detrend.R, R/mixedcirc_fitted.R
\name{mixedcirc_fitted}
\alias{mixedcirc_fitted}
\title{Extract the fitted circadian rhythm}
\usage{
mixedcirc_fitted(fit = NULL, per_group = FALSE, verbose = FALSE, ...)

mixedcirc_fitted(fit = NULL, per_group = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{fit}{an object of class mixedcirc_fit or mixedcirc_fit_list}

\item{per_group}{If TRUE, detrending is performed per group (default: FALSE)}

\item{verbose}{Show information about different stages of the processes. Default FALSE}

\item{...}{additionl arguments to the regression function}
}
\value{
A matrix of detrended data

A matrix of detrended data
}
\description{
This function extract the fitted values of circadian rhythm.

This function removes linear trend from the fitted values of circadian rhythm.
}
\details{
This method will first calculate the fitted values of the provided model followed by fitting a regression of form x~time. The residues of this regression model will be returned as the detrended data.

This method will first calculate the fitted values of the provided model followed by fitting a regression of form x~time. The residues of this regression model will be returned as the detrended data.
}
\examples{
library(mixedcirc)
data("circa_data")
results<-mixedcirc_detect(data_input = circa_data$data_matrix,
                         time = circa_data$time,group = circa_data$group,id = circa_data$id,period = 24,verbose = TRUE)
detrended<-mixedcirc_fitted(results)


library(mixedcirc)
data("circa_data")
results<-mixedcirc_detect(data_input = circa_data$data_matrix,
                         time = circa_data$time,group = circa_data$group,id = circa_data$id,period = 24,verbose = TRUE)
detrended<-mixedcirc_detrend(results)


}
