% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_fitted.R
\name{mixedcirc_fitted}
\alias{mixedcirc_fitted}
\title{Extract the fitted circadian rhythm}
\usage{
mixedcirc_fitted(fit = NULL, per_group = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{fit}{an object of class mixedcirc_fit or mixedcirc_fit_list}

\item{verbose}{Show information about different stages of the processes. Default FALSE}
}
\value{
A matrix of detrended data
}
\description{
This function extract the fitted values of circadian rhythm.
}
\details{

}
\examples{
library(mixedcirc)
data("circa_data")
results<-mixedcirc_detect(data_input = circa_data$data_matrix,
                         time = circa_data$time,group = circa_data$group,id = circa_data$id,period = 24,verbose = TRUE)
detrended<-mixedcirc_fitted(results)


}
