% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_integrate.R
\name{mixedcirc_getscore}
\alias{mixedcirc_getscore}
\title{Retrieved scores from mixedcirc_integration}
\usage{
mixedcirc_getscore(
  input = NULL,
  type = NULL,
  dataset = NULL,
  merge = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{input}{A mixedcirc_integration object}

\item{type}{what scores to retrieve (partial, average, loadings, model)}

\item{dataset}{name of the data set to extract the data from}

\item{merge}{If TRUE, the scores or loadings will be merged into a single data frame (default: FALSE). This has no effect if \code{type} is average!}

\item{verbose}{Enables printing logs}
}
\value{
A data frame of values
}
\description{
This functions performs circadian rhythm data integration using canonical correlations.
}
\details{
This method is used to extract scores and loadings from a mixedcirc_integration object
}
\examples{
library(mixedcirc)
data("circa_data")
data_input <- list(a=circa_data$data_matrix[,1:3],b=circa_data$data_matrix[,3:7])
results <- mixedcirc_integrate(data_input,time = circa_data$time,group = circa_data$group,id = circa_data$id)
score_matrix <- mixedcirc_getscore(results,type="average")
}
