% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcirc_interpolate.R
\name{mixedcirc_interpolate}
\alias{mixedcirc_interpolate}
\title{Generate circadian rhythm}
\usage{
mixedcirc_interpolate(
  x,
  period = 24,
  min_time = NULL,
  max_time = NULL,
  npoints = 200
)
}
\arguments{
\item{x}{A class of mixedcirc_fit}

\item{period}{The rhythm period. Default: 24}

\item{min_time}{Minimum time span to do the prediction. If NULL, it will be taken from the fit. Default: NULL}

\item{max_time}{Maximum time span to do the prediction. If NULL, it will be taken from the fit. Default: NULL}

\item{npoints}{Maximun number of data points to generate Default: 200}
}
\value{
A data.frame with all the elements of the model including "Y.hat" that is the interpolated data
}
\description{
This functions gets a mixedcirc_fit circadian rhythm fitted using mixedcirc_detect function and interpolate data points between two times.
}
\details{
In case of RRBS, the data is assumed to be log2. The M-values are calculated as Mathylated-Unmethylated
}
\examples{
data("circa_data")

results<-mixedcirc_detect(data_input = circa_data$data_matrix,
time = circa_data$time,group = circa_data$group,id = circa_data$id,period = 24,verbose = TRUE)
int_data<-mixedcirc_interpolate(results)
}
